% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.dass21.R
\name{dass21}
\alias{dass21}
\title{Calculate Depression Anxiety Stress Scales - 21 (DASS-21) scores}
\usage{
dass21(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Calculates DASS-21 (Lovibond & Lovibond, 1995) scores.  Expects a data frame
formatted by \code{\link[expfactory:process_expfactory_survey]{expfactory::process_expfactory_survey}}, containing
responses from a single participant. Participant id \emph{must} be in column \code{p}.
}
\examples{
\dontrun{
dass21_df <- expand.grid(token = participants$token, survey = 'dass-21-results.json') \%>\%
  rowwise() \%>\%
  do(., expfactory::process_expfactory_survey(.$token, paste('data/', .$token, '_finished/', .$survey, sep=''),
    flat=TRUE)) \%>\%
    rename(p = Token)
 process_dass21_data <- function(participant, df) {
   return(expfactory.dass21::dass21(df \%>\% filter(p == participant)))
 }
 dass21 <- dass21_df \%>\%
   rowwise() \%>\%
   do(., process_dass21_data(.$p, dass21_df)) \%>\%
   arrange(p)
}
}
\keyword{anxiety,}
\keyword{dass-21}
\keyword{depression,}
\keyword{stress,}
